# Introduce Log processor